backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
requests==2.31.0
beautifulsoup4==4.12.2
websockets==12.0
python-multipart==0.0.6
gTTS==2.5.1
pydantic==2.9.2
pydantic-core==2.23.4

openai==1.59.8
tiktoken==0.7.0

langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.9
langchain-openai==0.1.20
langsmith==0.1.147
chromadb==0.4.22
chromadb==0.4.22
langchain-chroma>=0.1.2,<0.2.0
langchain-text-splitters==0.2.4
